+++ db/migrate/20161222090545_add_pushed_by_to_versions.rb
@@ -1,0 +1,5 @@
+class AddPushedByToVersions < ActiveRecord::Migration
+  def change
+    add_column :versions, :pushed_by, :string
+  end
+end
