+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160929104437) do
+ActiveRecord::Schema.define(version: 20161222090545) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -177,6 +177,7 @@ ActiveRecord::Schema.define(version: 20160929104437) do
     t.string   "required_rubygems_version"
     t.string   "info_checksum"
     t.string   "yanked_info_checksum"
+    t.string   "pushed_by"
   end
 
   add_index "versions", ["built_at"], name: "index_versions_on_built_at", using: :btree
